+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.8] - 2020-01-23
+
+### Added
+- Option to build accessors for additional properties in generated `PHP` classes.
+- Option to declare default values for properties in generated `PHP` classes.
+
 ## [1.6.7] - 2020-01-04
 
 ### Added
@@ -43,6 +49,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Local file resolver in references.
 
+[1.6.8]: https://github.com/swaggest/json-cli/compare/v1.6.7...v1.6.8
 [1.6.7]: https://github.com/swaggest/json-cli/compare/v1.6.6...v1.6.7
 [1.6.6]: https://github.com/swaggest/json-cli/compare/v1.6.5...v1.6.6
 [1.6.5]: https://github.com/swaggest/json-cli/compare/v1.6.4...v1.6.5